Extract hash function into its own file
