+++ src/js/iqons.js
@@ -1,4 +1,5 @@
 // removeIf(production)
+import { makeHash } from './hash.js';
 import { hashToRGB } from './colors.js';
 // endRemoveIf(production)
 
@@ -7,7 +8,7 @@ export default class Iqons {
     /* Public API */
 
     static async svg(text) {
-        const hash = this._hash(text);
+        const hash = makeHash(text);
         return this._svgTemplate(
             hash[0], // color
             hash[2], // backgroundColor
@@ -128,48 +129,8 @@ ${ content }
         return index
     }
 
-    static _hash(text) {
-        const fullHash = ('' + text
-                .split('')
-                .map(c => Number(c.charCodeAt(0)) + 3)
-                .reduce((a, e) => a * (1 - a) * this.__chaosHash(e), 0.5))
-            .split('')
-            .reduce((a, e) => e + a, '');
-
-        const hash = fullHash
-            .replace('.', fullHash[5]) // Replace the dot as it cannot be parsed to int
-            .substr(4, 17);
-
-        // The index 5 of `fullHash` is currently unused (index 1 of `hash`,
-        // after cutting off the first 4 elements). Iqons.svg() is not using it.
-
-        // A small percentage of returned values are actually too short,
-        // leading to an invalid bottom index and feature color. Adding
-        // padding creates a bottom feature and accent color where no
-        // existed previously, thus it's not a disrupting change.
-        return Iqons._padEnd(hash, 13, fullHash[5]);
-    }
-
-    static __chaosHash(number) {
-        const k = 3.569956786876;
-        let a_n = 1 / number;
-        for (let i = 0; i < 100; i++) {
-            a_n = (1 - a_n) * a_n * k;
-        }
-        return a_n;
-    }
-
     static _getRandomId() {
         return Math.floor(Math.random() * 256);
-    }
-
-    // Polyfill
-    static _padEnd(string, maxLength, fillString) {
-        if (!!String.prototype.padEnd) return string.padEnd(maxLength, fillString);
-        else {
-            while (string.length < maxLength) string += fillString;
-            return string.substring(0, Math.max(string.length, maxLength));
-        }
     }
 }
